refactor: move search logic from compoents to custom hook
LP-14
